% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dendropulse.R
\name{get_dendropulse}
\alias{get_dendropulse}
\title{Connect to the DendroPulse PostgreSQL Database}
\usage{
get_dendropulse(admin = F)
}
\arguments{
\item{admin}{Logical.
\itemize{
\item \code{FALSE} (default): Connect using the read‑only credentials.
\item \code{TRUE}: Prompt interactively for an admin username and password.
}}
}
\value{
A \code{DBIConnection} object if the connection is successful.
Returns \code{NULL} and prints an error message if the connection fails.
}
\description{
This function creates a connection to the \emph{dendro_pulse} PostgreSQL database.
By default, it uses the read‑only "pulse" account.
If \code{admin = TRUE}, the function prompts the user to manually enter a username
and password via the console (safer for admin access).
}
\details{
\itemize{
\item Uses the \strong{RPostgres} driver to establish the connection.
}
}
\examples{
\dontrun{
# Connect with read‑only access
con <- get_dendropulse()

# Connect with admin credentials
con <- get_dendropulse(admin = TRUE)
}

}
